--- app/assets/stylesheets/application.scss
@@ -1,1 +1,0 @@
-@import "lib/govuk-frontend/govuk/base";
