+++ conf/app.routes
@@ -1,5 +1,5 @@
 # microservice specific routes
 
 ->         /hmrc-frontend           hmrcfrontend.Routes
-GET        /                        uk.gov.hmrc.jan24nicchangecalculatoradminfrontend.controllers.HelloWorldController.helloWorld
+GET        /                        controllers.HelloWorldController.helloWorld
 GET        /assets/*file            controllers.Assets.versioned(path = "/public", file: Asset)
