+++ conf/application.conf
@@ -23,10 +23,10 @@ play.filters.enabled += play.filters.csp.CSPFilter
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientModule"
 
 # Custom error handler
-play.http.errorHandler = "uk.gov.hmrc.jan24nicchangecalculatoradminfrontend.config.ErrorHandler"
+play.http.errorHandler = "config.ErrorHandler"
 
 # Play Modules
-play.modules.enabled += "uk.gov.hmrc.jan24nicchangecalculatoradminfrontend.config.Module"
+play.modules.enabled += "config.Module"
 
 microservice {
   services {
